+++ siman/small_functions.py
@@ -48,9 +48,9 @@ def is_list_like(obj):
 def is_string_like(s):
     return isinstance(s, string_types)
 
-def list2string(ilist):
+def list2string(ilist, joiner = ' '):
     #' '.join(['{:}']*len(lis)).format(*lis)
-    return ' '.join(np.array(ilist).astype(str))
+    return joiner.join(np.array(ilist).astype(str))
 
 
 def merge_dics(dic1, dic2):